Fix crash in trying manual search again before first search completes
Manual network search query and responses are not synchronous.
This leads to multiple responses being processed consecutively.
A dialog is cleared on receiving the first response
and subsequent responses through exception when attempting to dismiss
the same dialog. Ignore the exception thrown by
dialogs as it is not benign.

Change-Id: Icc092b093af04ddfb15b2e5b30ac123577eab05c
Crs-Fixed: 326736